Изменения от 09.06.2023

+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -3,6 +3,6 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.model.CommentEntity;
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
+public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
 
 }
